Introduce Bag Plot and HDR statistic filter.
Add a new type of plot to draw 'bag plots', a bivariate generalization
of the univariate boxplot.
Add a new statistic filter to compute the Highest Density Regions
that can be used to compute a density field used to draw the bag plot.

Add support for the case where the bag is reduced to 2 points.

Change-Id: I16852a46bacc3d74a72c001114cef9b287765c9d